+++ src/generated_utils/automaton_class.py
