+++ src/main.py
@@ -2,8 +2,8 @@ import sys
 import type_collector
 import type_builder
 import type_checker
-from lexer import tokenize
-from parser import parse
+from lexer_ply import tokenize
+from parser_ply import parse
 from translate_cool_cil import COOLToCILVisitor
 from translate_cil_mips import CILToMIPSVisitor
 from utils.utils import display_errors
